microservicioConsultas/requirements.txt
uvicorn==0.32.0
fastapi==0.115.0
openai==1.60.2
python-dotenv==0.20.0